requirements.txt
appdirs==1.4.3
kafka-python==1.3.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
